Format percentage

+++ Gaso/app/src/main/java/com/ericmguimaraes/gaso/model/FuzzyConsumption.java
@@ -105,7 +105,7 @@ public class FuzzyConsumption implements Serializable {
     }
 
     public float getPercentage(int value){
-        return ((int) ((float) value/getTotal()*100));
+        return ((float) value/getTotal()*100);
     }
 
 }
